primaholding-magiclink-app/requirements.txt
flask
pyjwt
flask-cors
